로그인 관련 권한 체크

+++ src/main/java/com/sbs/webp/lolHi/controller/usr/ArticleController.java
@@ -80,7 +80,20 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("/usr/article/doWrite")
-	public String doWrite(@RequestParam Map<String, Object> param, Model model) {
+	public String doWrite(HttpSession session, Model model, @RequestParam Map<String, Object> param) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int)session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("msg", "로그인 후 이용해주세요.");
+			model.addAttribute("replaceUri", "/usr/member/login");
+			
+			return "common/redirect";
+		}
+		
 		int id = articleService.writeArticle(param);
 		
 		model.addAttribute("msg", String.format("%d번 게시물이 생성되었습니다.", id));
@@ -90,7 +103,20 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("/usr/article/modify")
-	public String showModify(Model model, int id) {
+	public String showModify(HttpSession session, Model model, int id) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int)session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("msg", "로그인 후 이용해주세요.");
+			model.addAttribute("replaceUri", "/usr/member/login");
+			
+			return "common/redirect";
+		}
+		
 		Article article = articleService.getArticleById(id);
 		
 		model.addAttribute("article", article);
@@ -99,7 +125,20 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("/usr/article/doModify")
-	public String doModify(int id, String title, String body, Model model) {
+	public String doModify(HttpSession session, Model model, int id, String title, String body) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int)session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("msg", "로그인 후 이용해주세요.");
+			model.addAttribute("replaceUri", "/usr/member/login");
+			
+			return "common/redirect";
+		}
+		
 		articleService.modifyArticle(id, title, body);
 		
 		model.addAttribute("msg", String.format("%d번 게시물이 수정되었습니다.", id));
@@ -109,7 +148,20 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("/usr/article/doDelete")
-	public String doDelete(int id, Model model) {
+	public String doDelete(HttpSession session, Model model, int id) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int)session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("msg", "로그인 후 이용해주세요.");
+			model.addAttribute("replaceUri", "/usr/member/login");
+			
+			return "common/redirect";
+		}
+		
 		articleService.deleteArticleById(id);
 		
 		model.addAttribute("msg", String.format("%d번 게시물이 삭제되었습니다.", id));
